requirements.txt
fastapi[all]
pydantic
google-generativeai
python-dotenv